Fixed changes requested in PR review
