+++ .github/workflows/UploadFiles_build_and_test_on_main.yml
@@ -14,6 +14,6 @@ jobs:
     uses: FrendsPlatform/FrendsTasks/.github/workflows/linux_build_main.yml@main
     with:
       workdir: Frends.SFTP.UploadFiles
-      prebuild_command: docker-compose -f ./Frends.SFTP.UploadFiles.Tests/docker_compose.yml up -d && docker logs testikontti
+      prebuild_command: docker-compose -f ./Frends.SFTP.UploadFiles.Tests/ up -d
     secrets:
       badge_service_api_key: ${{ secrets.BADGE_SERVICE_API_KEY }}
