+++ .github/workflows/UploadFiles_build_and_test_on_push.yml
@@ -14,7 +14,7 @@ jobs:
     uses: FrendsPlatform/FrendsTasks/.github/workflows/linux_build_test.yml@main
     with:
       workdir: Frends.SFTP.UploadFiles
-      prebuild_command: docker-compose -f ./Frends.SFTP.UploadFiles.Tests/docker_compose.yml up -d
+      prebuild_command: docker-compose -f ./Frends.SFTP.UploadFiles.Tests/ up -d
     secrets:
       badge_service_api_key: ${{ secrets.BADGE_SERVICE_API_KEY }}
       test_feed_api_key: ${{ secrets.TASKS_TEST_FEED_API_KEY }}
