+++ Frends.SFTP.UploadFiles/Frends.SFTP.UploadFiles/Definitions/BatchContext.cs
@@ -1,35 +1,35 @@
 ﻿using System;
 using Renci.SshNet.Sftp;
 
-namespace Frends.SFTP.UploadFiles.Definitions
+namespace Frends.SFTP.UploadFiles.Definitions;
+
+/// <summary>
+/// BatchContext class for creating collection of Task parameters
+/// </summary>
+internal class BatchContext
 {
-    /// <summary>
-    /// BatchContext class for creating collection of Task parameters
-    /// </summary>
-    internal class BatchContext
-    {
-        public Info Info { get; set; }
+    public Info Info { get; set; }
 
-        public string TempWorkDir { get; set; }
+    public string TempWorkDir { get; set; }
 
-        public Options Options { get; set; }
+    public Options Options { get; set; }
 
-        public Guid InstanceId { get; set; }
+    public Guid InstanceId { get; set; }
 
-        public string ServiceId { get; set; }
+    public string ServiceId { get; set; }
 
-        public IEnumerable<FileItem> SourceFiles { get; set; }
+    public IEnumerable<FileItem> SourceFiles { get; set; }
 
-        public IEnumerable<SftpFile> DestinationFiles { get; set; }
+    public IEnumerable<SftpFile> DestinationFiles { get; set; }
 
-        public string RoutineUri { get; set; }
+    public string RoutineUri { get; set; }
 
-        public DateTime BatchTransferStartTime { get; set; }
+    public DateTime BatchTransferStartTime { get; set; }
 
-        public Source Source { get; set; }
+    public Source Source { get; set; }
 
-        public Destination Destination { get; set; }
+    public Destination Destination { get; set; }
 
-        public Connection Connection  { get; set; }
-    }
+    public Connection Connection  { get; set; }
 }
+
