+++ Frends.SFTP.UploadFiles/Frends.SFTP.UploadFiles/Definitions/FileItem.cs
@@ -1,42 +1,42 @@
 ﻿using Renci.SshNet.Sftp;
 
-namespace Frends.SFTP.UploadFiles.Definitions
+namespace Frends.SFTP.UploadFiles.Definitions;
+
+internal class FileItem
 {
-    internal class FileItem
+    public DateTime Modified { get; set; }
+
+    public string Name { get; set; }
+
+    public string NameWithMacrosExtended { get; set; }
+
+    public string FullPath { get; set; }
+
+    public long Size { get; set; }
+
+    public FileItem(SftpFile file)
     {
-        public DateTime Modified { get; set; }
-
-        public string Name { get; set; }
-
-        public string NameWithMacrosExtended { get; set; }
-
-        public string FullPath { get; set; }
-
-        public long Size { get; set; }
-
-        public FileItem(SftpFile file)
-        {
-            Modified = file.LastWriteTime;
-            Name = file.Name;
-            Size = file.Length;
-            FullPath = file.FullName;
-        }
-
-        public FileItem(string fullPath)
-        {
-            if (!File.Exists(fullPath))
-                throw new ArgumentException($"File does not exist: '{fullPath}");
-
-            var fi = new FileInfo(fullPath);
-            Modified = fi.LastWriteTime;
-            Name = Path.GetFileName(fullPath);
-            Size = fi.Length;
-            FullPath = fullPath;
-        }
-
-        /// <summary>
-        /// Default constructor, use only for testing.
-        /// </summary>
-        public FileItem() { }
+        Modified = file.LastWriteTime;
+        Name = file.Name;
+        Size = file.Length;
+        FullPath = file.FullName;
     }
+
+    public FileItem(string fullPath)
+    {
+        if (!File.Exists(fullPath))
+            throw new ArgumentException($"File does not exist: '{fullPath}");
+
+        var fi = new FileInfo(fullPath);
+        Modified = fi.LastWriteTime;
+        Name = Path.GetFileName(fullPath);
+        Size = fi.Length;
+        FullPath = fullPath;
+    }
+
+    /// <summary>
+    /// Default constructor, use only for testing.
+    /// </summary>
+    public FileItem() { }
 }
+
