+++ Frends.SFTP.UploadFiles/Frends.SFTP.UploadFiles/Definitions/FileTransferResult.cs
@@ -1,23 +1,23 @@
-﻿namespace Frends.SFTP.UploadFiles.Definitions
+﻿namespace Frends.SFTP.UploadFiles.Definitions;
+
+internal class FileTransferResult
 {
-    internal class FileTransferResult
-    {
-        public bool ActionSkipped { get; set; }
+    public bool ActionSkipped { get; set; }
 
-        public bool Success { get; set; }
+    public bool Success { get; set; }
 
-        public string UserResultMessage { get; set; }
+    public string UserResultMessage { get; set; }
 
-        public int SuccessfulTransferCount { get; set; }
+    public int SuccessfulTransferCount { get; set; }
 
-        public int FailedTransferCount { get; set; }
+    public int FailedTransferCount { get; set; }
 
-        public IEnumerable<string> TransferredFileNames { get; set; }
+    public IEnumerable<string> TransferredFileNames { get; set; }
 
-        public Dictionary<string, IList<string>> TransferErrors { get; set; }
+    public Dictionary<string, IList<string>> TransferErrors { get; set; }
 
-        public IEnumerable<string> TransferredFilePaths { get; set; }
+    public IEnumerable<string> TransferredFilePaths { get; set; }
 
-        public IDictionary<string, string> OperationsLog { get; set; }
-    }
+    public IDictionary<string, string> OperationsLog { get; set; }
 }
+
