+++ Frends.SFTP.UploadFiles/Frends.SFTP.UploadFiles/Definitions/SourceAction.cs
@@ -1,15 +1,15 @@
 ﻿// Pragma is for self-explanatory enum attributes.
 #pragma warning disable 1591
 
-namespace Frends.SFTP.UploadFiles.Definitions
+namespace Frends.SFTP.UploadFiles.Definitions;
+
+/// <summary>
+/// Enumeration to specify actions if the source file is not found.
+/// </summary>
+public enum SourceAction
 {
-    /// <summary>
-    /// Enumeration to specify actions if the source file is not found.
-    /// </summary>
-    public enum SourceAction
-    {
-        Error,
-        Info,
-        Ignore
-    }
+    Error,
+    Info,
+    Ignore
 }
+
