+++ Frends.SFTP.UploadFiles/Frends.SFTP.UploadFiles/Definitions/SourceOperation.cs
@@ -1,16 +1,16 @@
 ﻿// Pragma is for self-explanatory enum attributes.
 #pragma warning disable 1591
 
-namespace Frends.SFTP.UploadFiles.Definitions
+namespace Frends.SFTP.UploadFiles.Definitions;
+
+/// <summary>
+/// Enumeration to specify operation for the source file after transfer.
+/// </summary>
+public enum SourceOperation
 {
-    /// <summary>
-    /// Enumeration to specify operation for the source file after transfer.
-    /// </summary>
-    public enum SourceOperation
-    {
-        Delete,
-        Rename,
-        Move,
-        Nothing
-    }
+    Delete,
+    Rename,
+    Move,
+    Nothing
 }
+
