+++ Frends.SFTP.UploadFiles/Frends.SFTP.UploadFiles/Definitions/Util.cs
@@ -1,51 +1,51 @@
 ﻿using System.Text.RegularExpressions;
 
-namespace Frends.SFTP.UploadFiles.Definitions
+namespace Frends.SFTP.UploadFiles.Definitions;
+
+/// <summary>
+/// Helper methods for file modifications
+/// </summary>
+internal static class Util
 {
-    /// <summary>
-    /// Helper methods for file modifications
-    /// </summary>
-    internal static class Util
+    internal static string CreateUniqueFile(string toDir)
     {
-        internal static string CreateUniqueFile(string toDir)
+
+        if (!toDir.EndsWith("/") && !toDir.EndsWith("\\"))
         {
-
-            if (!toDir.EndsWith("/") && !toDir.EndsWith("\\"))
-            {
-                toDir = toDir + "\\";
-            }
-
-            return Path.GetFullPath(toDir + (DateTime.Now.Ticks + Path.GetRandomFileName()));
+            toDir = toDir + "\\";
         }
 
-        internal static string CreateUniqueFileName()
+        return Path.GetFullPath(toDir + (DateTime.Now.Ticks + Path.GetRandomFileName()));
+    }
+
+    internal static string CreateUniqueFileName()
+    {
+        return Path.ChangeExtension("frends_" + DateTime.Now.Ticks + Path.GetRandomFileName(), "8CO");
+    }
+
+    internal static bool FileMatchesMask(string filename, string mask)
+    {
+        const string regexEscape = "<regex>";
+        string pattern;
+
+        //check is pure regex wished to be used for matching
+        if (mask.StartsWith(regexEscape))
+            //use substring instead of string.replace just in case some has regex like '<regex>//File<regex>' or something else like that
+            pattern = mask.Substring(regexEscape.Length);
+        else
         {
-            return Path.ChangeExtension("frends_" + DateTime.Now.Ticks + Path.GetRandomFileName(), "8CO");
+            pattern = mask.Replace(".", "\\.");
+            pattern = pattern.Replace("*", ".*");
+            pattern = pattern.Replace("?", ".+");
+            pattern = String.Concat("^", pattern, "$");
         }
 
-        internal static bool FileMatchesMask(string filename, string mask)
-        {
-            const string regexEscape = "<regex>";
-            string pattern;
+        return Regex.IsMatch(filename, pattern, RegexOptions.IgnoreCase);
+    }
 
-            //check is pure regex wished to be used for matching
-            if (mask.StartsWith(regexEscape))
-                //use substring instead of string.replace just in case some has regex like '<regex>//File<regex>' or something else like that
-                pattern = mask.Substring(regexEscape.Length);
-            else
-            {
-                pattern = mask.Replace(".", "\\.");
-                pattern = pattern.Replace("*", ".*");
-                pattern = pattern.Replace("?", ".+");
-                pattern = String.Concat("^", pattern, "$");
-            }
-
-            return Regex.IsMatch(filename, pattern, RegexOptions.IgnoreCase);
-        }
-
-        internal static byte[] ConvertFingerprintToByteArray(string fingerprint)
-        {
-            return fingerprint.Split(':').Select(s => Convert.ToByte(s, 16)).ToArray();
-        }
+    internal static byte[] ConvertFingerprintToByteArray(string fingerprint)
+    {
+        return fingerprint.Split(':').Select(s => Convert.ToByte(s, 16)).ToArray();
     }
 }
+
